a11ybob.com/requirements.txt
Flask==3.1.*
pymongo[srv]==4.11.*
python-dotenv==1.1.*
